-- add function to perform separable aggregations in parts (still testing)
git-svn-id: https://codeforge.lbl.gov/anonscm/fastbit/trunk@438 9ec3b2dc-5744-0410-9993-b8723341f1db

+++ src/iroster.h
@@ -155,7 +155,7 @@ inline uint32_t ibis::roster::operator[](uint32_t i) const {
 	tmp = ind[i];
     }
     else if (inddes >= 0) {
-	if (i*sizeof(uint32_t) !=
+	if (static_cast<off_t>(i*sizeof(uint32_t)) !=
 	    UnixSeek(inddes, i*sizeof(uint32_t), SEEK_SET))
 	    return tmp;
 	if (sizeof(uint32_t) != UnixRead(inddes, &tmp, sizeof(uint32_t)))
